pyproject.toml
[project]
name = "gemini-watermark-remover"
version = "1.0.0"
description = "CLI tool to remove Gemini AI watermarks from images and videos"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["gemini", "watermark", "removal", "cli", "ffmpeg"]

dependencies = [
    "typer>=0.9.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "ffmpeg-python>=0.2.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
gemini-watermark-remover = "gemini_watermark_remover.cli:app"
gwr = "gemini_watermark_remover.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gemini_watermark_remover"]

[tool.ruff]
line-length = 100
target-version = "py310"
